+++ scripts/preprocessing/_03_make_epochs.py
@@ -90,7 +90,9 @@ def run_epochs(*, cfg, subject, session=None):
             decim=cfg.decim
         )
 
-        # Only keep epochs that will be analyzed
+        # Only keep epochs of conditions that will be analyzed. Removal of
+        # epochs based on metadata queries will take place later to reduce the
+        # amount of logging output.
         if cfg.task != 'rest':
             if isinstance(cfg.conditions, dict):
                 conditions = list(cfg.conditions.keys())
@@ -124,7 +126,30 @@ def run_epochs(*, cfg, subject, session=None):
     logger.info(**gen_log_kwargs(message=msg, subject=subject,
                                  session=session))
 
-    msg = 'Writing epochs to disk'
+    # Lastly, exclude epochs based on metadata.
+    if cfg.epochs_metadata_excludes:
+        msg = f'Excluding epochs based on metadata queries …'
+        logger.info(**gen_log_kwargs(message=msg, subject=subject,
+                                     session=session))
+
+    for exclude in cfg.epochs_metadata_excludes:
+        n_epochs_before = len(epochs)
+        try:
+            epochs = epochs[f'not ({exclude})']
+        except KeyError:
+            msg = f'Metadata query failed to select any columns: {exclude}'
+            logger.warn(**gen_log_kwargs(message=msg, subject=subject,
+                                         session=session))
+            continue
+
+        n_epochs_after = len(epochs)
+        n_epochs_diff = n_epochs_before - n_epochs_after
+        msg = (f'Removed {n_epochs_diff} epochs matching metadata query: '
+                f'{exclude}')
+        logger.info(**gen_log_kwargs(message=msg, subject=subject,
+                                     session=session))
+
+    msg = f'Writing {len(epochs)} epochs to disk.'
     logger.info(**gen_log_kwargs(message=msg, subject=subject,
                                  session=session))
     epochs_fname = bids_path.copy().update(suffix='epo', check=False)
@@ -160,6 +185,7 @@ def get_config(
         epochs_metadata_tmax=config.epochs_metadata_tmax,
         epochs_metadata_keep_first=config.epochs_metadata_keep_first,
         epochs_metadata_keep_last=config.epochs_metadata_keep_last,
+        epochs_metadata_excludes=config.epochs_metadata_excludes,
         event_repeated=config.event_repeated,
         decim=config.decim,
         ch_types=config.ch_types,